Move non-repudiation in-memory backend to testing module
This will allow to share it with the non-repudiation API tests.

changelog_begin
changelog_end
