+++ runtime-components/non-repudiation/BUILD.bazel
@@ -45,6 +45,7 @@ da_scala_test(
         "//libs-scala/grpc-test-utils",
         "//libs-scala/resources",
         "//runtime-components/non-repudiation-client",
+        "//runtime-components/non-repudiation-testing",
         "@maven//:com_google_protobuf_protobuf_java",
         "@maven//:io_grpc_grpc_api",
         "@maven//:io_grpc_grpc_core",
