+++ runtime-components/non-repudiation/src/test/scala/com/daml/nonrepudiation/NonRepudiationProxySpec.scala
@@ -5,7 +5,7 @@ package com.daml.nonrepudiation
 
 import java.security.PrivateKey
 import java.security.cert.X509Certificate
-import java.time.{Clock, Duration, Instant, ZoneId}
+import java.time.{Clock, Instant, ZoneId}
 
 import com.daml.grpc.test.GrpcServer
 import com.daml.nonrepudiation.SignedPayloadRepository.KeyEncoder
@@ -15,10 +15,6 @@ import io.grpc.{Channel, StatusRuntimeException}
 import org.scalatest.Inside
 import org.scalatest.flatspec.AsyncFlatSpec
 import org.scalatest.matchers.should.Matchers
-import sun.security.tools.keytool.CertAndKeyGen
-import sun.security.x509.X500Name
-
-import scala.collection.concurrent.TrieMap
 
 final class NonRepudiationProxySpec
     extends AsyncFlatSpec
@@ -124,7 +120,7 @@ final class NonRepudiationProxySpec
     val Setup(certificates, signatures, _, _, proxyBuilder, proxyChannel) =
       Setup.newInstance[String]
 
-    val (privateKey, certificate) = Setup.generateKeyAndCertificate()
+    val (privateKey, certificate) = testing.generateKeyAndCertificate()
 
     NonRepudiationProxy
       .owner(
@@ -160,52 +156,17 @@ object NonRepudiationProxySpec {
 
   object Setup {
 
-    def generateKeyAndCertificate(): (PrivateKey, X509Certificate) = {
-      val generator = new CertAndKeyGen(AlgorithmString.RSA, AlgorithmString.SHA256withRSA)
-      generator.generate(2048)
-      val privateKey = generator.getPrivateKey
-      val certificate = generator.getSelfCertificate(
-        new X500Name("CN=Non-Repudiation Test,O=Digital Asset,L=Zurich,C=CH"),
-        Duration.ofHours(1).getSeconds,
-      )
-      (privateKey, certificate)
-    }
-
     def newInstance[Key: KeyEncoder]: Setup[Key] = {
-      val certificates = new Certificates
-      val signatures = new SignedPayloads
+      val certificates = new testing.Certificates
+      val signatures = new testing.SignedPayloads
       val proxyName = InProcessServerBuilder.generateName()
       val proxyBuilder = InProcessServerBuilder.forName(proxyName)
       val proxyChannel = InProcessChannelBuilder.forName(proxyName).build()
-      val (privateKey, certificate) = generateKeyAndCertificate()
+      val (privateKey, certificate) = testing.generateKeyAndCertificate()
       certificates.put(certificate)
       Setup(certificates, signatures, privateKey, certificate, proxyBuilder, proxyChannel)
     }
 
-  }
-
-  final class SignedPayloads[Key: KeyEncoder] extends SignedPayloadRepository[Key] {
-    private val map = TrieMap.empty[Key, SignedPayload]
-    override def put(signedPayload: SignedPayload): Unit = {
-      val _ = map.put(keyEncoder.encode(signedPayload.payload), signedPayload)
-    }
-
-    override def get(key: Key): Iterable[SignedPayload] =
-      map.get(key).toList
-  }
-
-  final class Certificates extends CertificateRepository {
-
-    private val map = TrieMap.empty[FingerprintBytes, X509Certificate]
-
-    override def get(fingerprint: FingerprintBytes): Option[X509Certificate] =
-      map.get(fingerprint)
-
-    override def put(certificate: X509Certificate): FingerprintBytes = {
-      val fingerprint = FingerprintBytes.compute(certificate)
-      map.put(fingerprint, certificate)
-      fingerprint
-    }
   }
 
 }
